\name{country.names}
\alias{country.names}
\docType{data}
\title{
  Country Names
}
\description{
  A list containing the country names used by various organizations, including 
  the International Monetary Fund (IMF), International Standards Organization (ISO),
  Penn World Tables (PWT), World Bank (WB), and World Health Organization (WHO).
  The default names are mostly derived from the ISO name set.
}
\usage{data(country.names)}
\format{
  The format is:
List of 6
 $ default: chr [1:260] "Afghanistan" "Aland Islands" "Albania" "Algeria" ...
 $ imf    : chr [1:184] "Islamic Republic of Afghanistan" "Albania" "Algeria" "Angola" ...
 $ iso    : chr [1:249] "Afghanistan" "Aland Islands" "Albania" "Algeria" ...
 $ pwt    : chr [1:189] "Afghanistan" "Albania" "Algeria" "Angola" ...
 $ wb     : chr [1:217] "Afghanistan" "Albania" "Algeria" "American Samoa" ...
 $ who    : chr [1:193] "Afghanistan" "Albania" "Algeria" "Andorra" ...
}
\examples{
data(country.names)
}
\keyword{datasets}
